TRS: Lexicon + RewriteSystem API
