+++ src/trs/mod.rs
@@ -1,23 +1,95 @@
 //! (representation) Polymorphically-typed term rewriting system.
+//!
+//! An evaluatable first-order [Term Rewriting System][0] (TRS) with a [Hindley-Milner type
+//! system][1].
+//!
+//! [0]: https://wikipedia.org/wiki/Hindley–Milner_type_system
+//!      "Wikipedia - Hindley-Milner Type System"
+//! [1]: https://en.wikipedia.org/wiki/Rewriting#Term_rewriting_systems
+//!      "Wikipedia - Term Rewriting Systems"
 
+mod lexicon;
+mod rewrite;
 mod trace;
-mod trs;
-pub use self::trs::{SampleError, TypeError, TRS, TRSParams, TRSSpace};
+pub use self::lexicon::{GeneticParams, Lexicon};
+pub use self::rewrite::{RewriteStrategy, RewriteSystem};
 use Task;
 
-use polytype::TypeSchema;
+use polytype;
+use std::fmt;
 use term_rewriting::Rule;
+
+#[derive(Debug, Clone)]
+pub enum TypeError {
+    Unification(polytype::UnificationError),
+    OpNotFound,
+    VarNotFound,
+}
+impl From<polytype::UnificationError> for TypeError {
+    fn from(e: polytype::UnificationError) -> TypeError {
+        TypeError::Unification(e)
+    }
+}
+impl fmt::Display for TypeError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match *self {
+            TypeError::Unification(ref e) => write!(f, "unification error: {}", e),
+            TypeError::OpNotFound => write!(f, "operator not found"),
+            TypeError::VarNotFound => write!(f, "variable not found"),
+        }
+    }
+}
+impl ::std::error::Error for TypeError {
+    fn description(&self) -> &'static str {
+        "type error"
+    }
+}
+
+#[derive(Debug, Clone)]
+pub enum SampleError {
+    TypeError(TypeError),
+    DepthExceeded(usize, usize),
+    OptionsExhausted,
+    Subterm,
+}
+impl From<TypeError> for SampleError {
+    fn from(e: TypeError) -> SampleError {
+        SampleError::TypeError(e)
+    }
+}
+impl From<polytype::UnificationError> for SampleError {
+    fn from(e: polytype::UnificationError) -> SampleError {
+        SampleError::TypeError(TypeError::Unification(e))
+    }
+}
+impl fmt::Display for SampleError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match *self {
+            SampleError::TypeError(ref e) => write!(f, "type error: {}", e),
+            SampleError::DepthExceeded(depth, max_depth) => {
+                write!(f, "depth {} exceeded maximum of {}", depth, max_depth)
+            }
+            SampleError::OptionsExhausted => write!(f, "failed to sample (options exhausted)"),
+            SampleError::Subterm => write!(f, "cannot sample subterm"),
+        }
+    }
+}
+impl ::std::error::Error for SampleError {
+    fn description(&self) -> &'static str {
+        "sample error"
+    }
+}
 
 pub fn make_task_from_data(
     data: &[Rule],
-    tp: TypeSchema,
+    tp: polytype::TypeSchema,
     p_partial: f64,
     temperature: f64,
     prior_temperature: f64,
     ll_temperature: f64,
-) -> Task<TRSSpace, TRS, ()> {
+) -> Task<Lexicon, RewriteSystem, ()> {
     Task {
-        oracle: Box::new(move |_s: &TRSSpace, h: &TRS| {
+        oracle: Box::new(move |_s: &Lexicon, h: &RewriteSystem| {
             // TODO: only getting information from temperature-adjusted evaluation
             h.posterior(
                 data,
